Map view for albums in galleria theme
This feature hacks galleria to show an openlayers map instead of the
picture. The pictures are represented by dots on the maps and clicking
them or the galleria thumbnails update the other (and the selected
picture).

By default, the map isn't shown: it needs to be enabled by the show_map
options.